run models implementation but still with bugs
